refactor(galaxian): extract cclimber_audio device + relocate harem protection
The shared galaxian machine/sound code (gal_run.cpp, gal_sound.cpp) called a
sample player and protection routines that were defined in the d_galaxian.cpp
game driver, so a single galaxian driver (e.g. d_119) could not link without
pulling in the whole game driver.

- Crazy Climber / Moon Shuttle sample player -> new cclimber_audio.cpp, a
  standalone unit mirroring cclimber_audio device, added to the galaxian
  group_srcs. The game driver drives it through a small init/exit/setter API in
  gal.h (cclimber_sample_init/exit/rom/w_freq/w_vol/start).
- Harem bootleg protection (opcode/data bankswitch + shift-register decrypt) ->
  gal_run.cpp, the shared machine file that already dispatches it; it uses only
  shared galaxian globals. The driver calls harem_bankswitch/harem_decrypt_scan.

d_119 now links standalone; full and galaxian-group builds link; all 649
drivers unchanged and statically complete. Logic relocated verbatim (behavior
preserved). Synced makefile .o and VS project files (fba_qt.pro globs the dir).
