+++ projectfiles/visualstudio-2003-libretro-xbox1/fba_vs2003_libretro_xbox1.vcproj
@@ -539,6 +539,9 @@
 							RelativePath="..\..\src\burn\drv\galaxian\gal_gfx.cpp">
 						</File>
 						<File
+							RelativePath="..\..\src\burn\drv\galaxian\cclimber_audio.cpp">
+						</File>
+						<File
 							RelativePath="..\..\src\burn\drv\galaxian\gal_run.cpp">
 						</File>
 						<File
