+++ src/burn/drv/galaxian/gal.h
@@ -288,6 +288,8 @@ void MapHarem();
 void MapNamenayo();
 void MapTurtles();
 void MapScobra();
+void harem_bankswitch(INT32 bank);
+void harem_decrypt_scan(INT32 nAction);
 void harem_decrypt_bit_write(UINT8 data);
 void harem_decrypt_clk_write(UINT8 data);
 void harem_decrypt_rst_write(UINT8 data);
@@ -341,6 +343,13 @@ void GalRenderSoundSamples(INT16 *pSoundBuf, INT32 nLength);
 void GalaxianSoundWrite(UINT32 Offset, UINT8 d);
 void GalaxianLfoFreqWrite(UINT32 Offset, UINT8 d);
 void GalaxianSoundUpdateTimers();
+// Crazy Climber / Moon Shuttle sample player (cclimber_audio.cpp)
+void cclimber_sample_init();
+void cclimber_sample_exit();
+UINT8 *cclimber_sample_rom();
+void cclimber_sample_w_freq(UINT8 data);
+void cclimber_sample_w_vol(UINT8 data);
+void cclimber_sample_start();
 void cclimber_sample_render(INT16 *buffer, INT32 nLen);
 void cclimber_sample_scan();
 void cclimber_sample_num(UINT32, UINT32 data);
